Add request()->validate() macro
This adds a new `validate()` macro to `\Next\Http\Request`.

The macro will automatically run the validation against the request when
it is called (no need to call `$validation->validate()`). It will also
infer what request parameters to validate by reading the keys from your
rules.

If you want to use different error messages to the default then you may
specify them as a second argument to `validate()`.

The validation macro is registered within Next\Validation\Proxy so that
if you remove the validation proxy from your application config then the
validation will cease to function. There will be no need to check
whether there is a validation instance available, like you would have to
do if it was a dedicated method directly on the `Next\Http\Request`
class.