Create Next\Console\Application and Next\Console\Command
The new \Next\Console\Application extends the Symfony console
application so it can be a drop-in replacement to the current console.

Application will automatically register the Collision error handler, so
that can be removed from the application. Also, the Application will set
an instance of \Next\App on any console command that extends
\Next\Console\Command.

You don't have to extend \Next\Console\Command so we can still utilise
3rd party commands extending \Symfony\Components\Console\Command\Command.

If your command extends \Next\Console\Command you must define a
`handle()` method instead of using the usual `execute()` method you
would with a Symfony command. This is because we have overwritten the
`execute()` command to utilise the `\Next\App` instance to inject any
dependencies specified in the `handle()` method signature. For example:

```
public function handle(LoggerInterface $logger) {
    // Interact with the logger
}
```

You may be wondering how you will interact with the input and output if
they're not passed through the handle method, as that is what you're
used to if you're familiar with Symfony commands. Well, there are now
some nice methods on the command for interacting with IO.

- `$this->argument('name');`
- `$this->option('name');`
- `$this->ask('question');`
- `$this->confirm('question');`
- `$this->info('message');`
- `$this->comment('comment');`
- `$this->warn('message');`
- etc.

If there is anything that isn't covered by the new IO API then you have
a couple new properties available to you `$this->input` and
`$this->output`.