Ensure Cache::remember only calls closure once
The initial intention for Cache::remember was to be an alias to
Cache::store that accepted a closure. However, after talking to
@assertchris we agreed that it should return the cached value instead of
calling the closure each time.

In order to get the test to pass I had to change the type of `$factory`
from `\Closure` to `callable`. This only adds support for any class that
defines an `__invoke` method and still supports regular `Closure`s so
this change should be okay.